backend/requirements.txt
bcrypt==4.0.1
passlib==1.7.4
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.17.0
cryptography==44.0.1
db-dtypes==1.4.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0
fastapi==0.115.10
Flask==3.0.3
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.30.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
PyJWT==2.10.1
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.63.0
packaging==24.2
pandas==2.2.2
pandas-gbq==0.23.1
pg8000==1.31.2
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
scramp==1.4.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.0.3
yarl==1.18.3
zipp==3.21.0
python-jose[cryptography]
schedule
asyncio
pandas
beautifulsoup4
playwright
python-dotenv==1.0.1
